added common api in store
